backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
sqlalchemy==2.0.37
pydantic==2.10.4
requests==2.32.3
Pillow==11.2.1
